Set no_proxy env var from NO_PROXY
The new library we use for making a websocket connection to cattle
looks for the no_proxy environment variable instead of NO_PROXY. So,
set the lowercase version based off the value of the uppercase version.
Note that the uppercase version must still be set because we still use
the library that looks for it.