Add support to download and run local agents
To support networks that may not have access to CDN the agents
are now downloaded into the rancher/server container. They are
compressed and accessed through the CATTLE_ settings. The default
behavior will now be to use the local agents. If a user wants to
leverage the CDN versions a `-e USE_LOCAL_AGENTS=false` must be
passed when starting the container.