Always set token cookie as secure
Previous method for deciding whether the request was secure (and
setting the secure flag on the cookie accordingly) doesn't work
because the URL on the request is relative and thus the scheme is
not set.

Rancher will always have TLS enabled, so always setting the secure
cookie flag to true is acceptable.