add timeout to 60 seconds
the while loop launched by rancher-agent-upgrade check the image every 5
seconds to see if agent images has been upgraded. There is also a race
condition that the rancher-agent-upgrade takes more than 5 seconds to
finish then we end up with a endless loop. So add timeout to 60 seconds
to give enough time for agent to launch.