Integrate princiapsl API with PRTB/CRTB
Allow users to assign group and user principal IDs to PRTBs/CRTBs
Also, some refactoring to minimize duplicate code and address a
bug found along the way. The bug was that the mgmt plane and the
user plan were keying off the same label value "owner" label to
determine when to created or delete bindings. This caused the logic
to fail in the "local" cluster that is also the mgmt plane and
erroneously delete bindings.