Set display name annotation on PRTB/CRTB
If a PRTB/CRTB is created with the userPrincipalId set, we will
ultimately need to ensure that the user exists in rancher. This
is done in a controller, which does not have access to the principals
API. So, set the principal's display name as an annotation on the
RTB during the API request (where we do have access to the principals
API) so that it can be used to ensure the user in the controller.