Cleaning up user bindings
Adds a user lifecycle to properly clean up cluster, global, and
project role bindings after a user is deleted from rancher.  This is to
prevent a user from still having access to a k8s cluster after
they have been deleted.

Adds retry logic to user_store so that the system will retry
creating a user if it fails due to the user lifecycle updating the user
with annotations after it is created.

Fixed a bug where the admin user would fail to be created on
cluster startup.

Issue:
https://github.com/rancher/rancher/issues/11378