Using pod security policy template project bindings
Pod security policy templates (pspt) were previous set on the project
resource.  This allowed anyone with project owner permissions to modify
the pspt attached to a project, which is not desired.  This change makes
pod security policy templates set in the pod security policy template
project binding resource, so that separate permission can be set for it.

Updated all existing handlers to accomodate the new change.

Also fixed an issue where cluster errors appeared in the logs after
giving a cluster a default PSPT.

Issue:
https://github.com/rancher/rancher/issues/12049

Fix repeated cluster errors in logs

Errors due to not being able to find a PSPT appear in the logs
repeatedly.

Issue:
https://github.com/rancher/rancher/issues/11773