Fix auth for apps, secrets, and volumes
Apps permissions need to be in roles that can perform workload operations.
Since app lives in the project namespace on the mgmt plane, it is also
added to the list of resources whose rules should be reflected in the
management plane project namespace.

Secrets permissions need to be reflected in the management plane project
namespace.

Volumes roles need to have consistent access to storageclasses and
persistentvolumes.