Clean up user token after deletion
Currently when a user is deleted from Rancher, their user token is not
being deleted.  This manifests as a unclear error message when the user
attempts to modify or query the k8s cluster via kubectl.

This changes the user lfc to delete the token while the other user
bindings are being deleted.

Issue:
https://github.com/rancher/rancher/issues/12287