Fix PSP race condition
Sometimes when a role binding is deleted it fails show when the delete is
followed by an immediate .Get.  This causes the default cluster PSP to
sometimes fail to create within a namespace.

This change adds additional checks before a role binding is deleted to
prevent the default PSPT role binding from being deleted unnecessarily.

Issue:
https://github.com/rancher/rancher/issues/12342