Fix login error
There are now two places that set the local principal on a user:
a controller and the api handler for when a user first logs in.
This leads to a scenario where the api creates the user, but the
controller is first to update the user with the principal and then
the api handler tries and gets a conflict error. This causes the
login to fail.

Solution is to remove the logic from the api handler so that it
is only performed in the controller.