Fix role creation for mgmt plane resources
- An annotation is set on the cluster when the initial creator role is
populated, but the logic was setting the annotation regardless of whether
or not the roles were populated without error. Changed this such that we
check for an error before setting the annotation on the cluster.

- The logic that created/updated the role was updating the role in a tight
loop and erroring out because of verion conflicts. Changed this such that
the role is only updated once after the loop is exited.