Add retry logic for creating/updating userAttr
If multiple login requests for the same user come in simultaneously, its
possbile to get a conflict on creating or updating the userAttribute
resource. This fix mitigates that by creating it in a backoff loop and if
we get an "already exists" error when trying to create, attenot a get
and update instead.