Fix logic around restricted access
Restricted access means that in addition to the explicitly whitelisted
set of user and groups that can log in, a user who is a meber of any
project or cluster can log in.

To accomplish the second half, we check if the user's individual and group
principals exist on any project or cluster RoleTemplateBindings (PRTB, CRTB).
However, the existing logic does not account for the CRTB/PRTB that the backend
creates for the owner of the project or cluster. That CRTB/PRTB does not
have a user or group principal field set; it just has the UserName (aka the
user ID) field set.

This change fixes the issue by also considering the user's name when determining
if the user is a member of a cluster or project.