Use proper index when determing if binding exists
When resolving roleBindings for a project, we check to see if a
binding exists. If none exists, then we create one. However, this
code has two indexes: one for clusterRoleBindings and one for
roleBindings. This code was incorrectly using the clusterRoleBinding
index while operating on roleBindings. This resulted in the controller
never finding the roleBinding and thus ALWAYS creating a new roleBinding,
which results in many duplicate roleBindings existing.