Fix subject comparison on clusterRoleBinding
When the controller for a globalRoleBinding runs, we determine if we
need to create, update, or delete a corresponding clusterRoleBinding.

For update, if we find an existing clusterRoleBinding, we then do a
deepEquals on the binding's actual subject (that represents the user)
against the expected subject (constructed from the globalRoleBinding).

This comparison was flawed because we were not setting the APIGroup
field and so we were always performing an update on the clusterRoleBinding.