Use logrus for logging http server panics
http.Server will recover from panics and log the panic's goroutine
dump to its ErrorLogger. By default, the error logger uses the
log package's builtin logger, which typically just logs to stderr.
However, when running rancher in "embedded" mode, kubernetes code
runs which reconfigures the builtin logger to use glog. In this case,
glog is just writing to a file in the /tmp dir, which was causing us
to not print panics.