Update isEqual method to handle nil empty maps as the same
The node objects were constantly being updated when one of the objects
had a an empty map {} vs a nil map. This comparison takes that into account
and treats them the same. Also, if the objects are not equal a debug
statement was added to see which comparison test was false.