Fix kubeconfig token race condtion
If multiple calls to generateKubeConfig action are made near simultaneously,
we can hit a race condition where the second call fails to find the token
in the cache and also fails to create it. Fix is to get the token
from the API if the create call fails with an "already exists" error.

This was manifesting itself in our drone builds where our tests are run
in parallel.