StatusEqual check fix
Using DeepEquals on the NodeStatus fields resulted in unneeded object
updates. The list of images being compared contained the same content
but a different order which were seen as unequal. As far as Rancher is
concerned, the same content was equal. So a separate method is added
breaking out the fields into their own tests.

Also, since Images is not used at this time, we can skip it.