Add user enabled check for saml
Problem:
When attempting to login with a user which is not enabled no error is
presented in the UI.

Solution:
Add a check on the user for enabled and redirect with an error.
Previously this was being caught on the token so the UI would not
show/redirect an error.