Fix an issue where members could not be resolved
Problem: The UI cannot display project and cluster creator
on the members page. This is because we change the permissions
on the user role such that non-admin users can no longer list
users at /v3/users.

Solution: The UI only uses /v3/users to look up members as a
fallback. It first tries to hit /v3/principals, which normal users
DO have access to. The problem was that we were not setting the
userPrincipalId on the special "creator" CRTB/PRTBs that are
created when a project or clsuter is created.

So, we just need to set that userPrincipalId field on the CRTBs
and PRTBs to allow the UI to resolve the principal properly.