Adding controller to add a temporary security token annotation
This change adds a controller that looks for the `sessionToken` to be present
on the EKS config, and if it is there, add an annotation to the cluster.  The
reason this is necessary is becuase credentials that have a `sessionToken`
will expire and the UI must take additional steps to refresh the credentials
before attempting an operation on the cluster, but because `sessionToken` is a
password field, it will be hidden from the UI also.

Issue:
https://github.com/rancher/rancher/issues/16005