Fix snapshot upgrade
For existing clusters, the snapshot value for the etcd service is nil. The RKE driver will interpret this as true, but we want it to behave as false in the API. We catch it before norman applies defaults (which would cause it to show as true), and say that if it is nil, it is false.

We also now convert a nil value for the snapshot setting as false. This only happens when we're already planning on passing the clusterspec to the driver update. This way, nil snapshot values which are considered false in the UI don't get handled as true in the RKE driver