Fixing missing config on list
This change fixes a bug where the config fields for clusters backed by
managed kubernetes providers were not displaying their configuration correctly
when listed.

This is accomplished by moving the logic that transposes the generic engine
config field to the provider specific config field to the transformer func for
the cluster store.  It also removes now unnecessary transpose calls in the
various `ByID`, `Create`, and `Update` funcs and also removes unnecessary
logic from the formatter.

Issue:
https://github.com/rancher/rancher/issues/16813