Fix alias type dns record issue
Problem:

As Alias type DNS record is a service without selector. Kubernetes endpoint controller will not create the endpoints automatically for it. Rancher is responsible for creating the endpoints. However the name field of the endpoint port is wrong. It should be the same with the name defined in service port.

Solution:

Fix it by always using the ports defined in service spec

https://github.com/rancher/rancher/issues/17321