Giving users the ability to view kontainer drivers
This change gives standard users the ability to view kontainer driver
resources.  This is important because in order to create a cluster the user
must be able to list kontainer drivers to know what options are available.
This capability is also added to the `Create Clusters` role as they also need
to be able to view kontainer drivers.

This also adds a `Manage Cluster Drivers` role to mirror the `Manage Node
Drivers` role.

Issue:
https://github.com/rancher/rancher/issues/17634

Issue: