Auth refresh in HA unreliable
Problem: In an HA setup, the user auth refresh logic that is triggered
from an API action did not work properly. The controllers and logic
were setup such that it would only work on the "leader" node. On
non-leader nodes, the logic would panic.

This fixes the problem by move the logic to the proper API scope,
ensuring that it is available and proeprly configured on every node.