Make cacerts setting read-only
**Problem:**
The contents of `/v3/cacerts` should not be manually edited.
It's either set to the value of the self signed generated
certificate, or the bind mounted certificate, or set to empty
when `--no-cacerts` is passed. Editing the value manually is
not supported and cause issues.

**Solution:**
The `cacerts` settings can not be updated.