Fix panic in refresher logic
In a previous commit, we fix a problem where the refresher would not
work properly in HA setups because it was only running properly on the
leader node. We fixed that by confiugring it to run on all nodes
and by supplying a reference to a refresher to the custom api action
logic.

However, we still have a global scoped refresher that runs on the leader.
This is because it needs launched from a cron. In the previous fix,
we missed one reference to the global variable. It should have been changed
to the local function receiver varaible, but wasn't.

This commit fixes that issue by changing that reference and avoids future issues
like that by separating all the globally scoped references and logic into their
own file (daemon.go) and all the logic that is scoped solely the refresher itself
into another file (refresher.go).