Resolve the confusion that occurs when disabling monitoring
**Problem:**
Assuming there are two clusters, monitoring is enabled. When disabling
the monitoring of cluster A, monitoring app of cluster B also be
dropped. But cluster B is still in monitoring `enabling` statue.

**Solution:**
When withdrawing the monitoring app, we need to increase a `clusterID`
parameter to ensure dropping the app of expected cluster.

**Issue:**
https://github.com/rancher/rancher/issues/18146