Adding code to address aks timeout issue:
This change addresses an issue where AKS clusters would timeout due to Azure
lbs failing to send tcpResets (RST) after an idle timeout.  This causes errors
to appear on the Rancher UI and logs.  This issue resolves that by adding a
keep alive to the dialer and changing some timeouts to prevent the errors from
appearing.

This is only for AKS cluster provisioned through the AKS Kontainer Driver
integration, imported clusters still exhibit this behavior.

Issue:
https://github.com/rancher/rancher/issues/14354