Ensure auth provider can find multiclusterapp member's principal
A user could have turned on external auth and added some user from that external auth
provider as a member of a multiclusterapp. After this when user logs back in as local
user, searching and adding new members from the external auth provider to the existing
multiclusterapp will fail. But even upgrading the multiclusterapp is failing, because
during upgrade, there is a call to getPrincipal of the members and set their displayNames
This call can be skipped if the current auth provider found in token is for local auth,
whereas the principalIDs of the members are from an external auth provider