Ensure temp credential annotation is set correctly
This change ensures that the temporary security credentials annotation is set
properly on an EKS cluster that has a session token.  Previously if the
session token field was missing or nil it would fail to be set in certain
places.  this change ensures that the temp security creds annotation will only
be set to false if the field is neither missing nor empty.

Issue:
https://github.com/rancher/rancher/issues/18285