don't attempt to deploy app in a terminating namespace
Issue: If mcApp is deleted and immediately created in same project,
there's a chance app namespace is still getting terminated.
We'd still be able to fetch it, but if it gets deleted before helm
checks, it'll create a new ns which won't have required labels.
It doesn't always reproduce, but better to add a check.