Fix for cluster driver nil error
Problem: A cluster driver could be deleted when there were active
clusters that depended on that driver.

Solution: The API will not crash if a cluster driver is unable to be
removed. Kontainer drivers are no longer able to be removed via the UI
or API if there is a cluster that uses that driver.