Add project monitoring config refresher
**Problem:**
The project monitoring can scrape data from ns which is already moved
to another project. The prometheus operator does not refesh the
prometheus configuration after the ns label is changed.

**Solution:**
Set an annotation which contains project ns list for prometheus obj.
And update the annotation when namespace projectID label is changed.