remove resourceQuota cleanup controller
GKE started adding resource_quota by default to cluster with nodes<5,
which cannot be deleted. In presence of more than one resource quotas,
the smaller one is always respected. So it's safe to remove the
cleanup controller for everything, with the assumption that if
resource quotas created through kubectl are present in addition to
the one created by us, k8s will respect the more restrictive one.