Fix kontainer-engine driver reactivation
Problem:
If you deactivate and then reactivate a kontainer-engine driver,
the schema for it will not be readded. This is because there is logic
that only creates/updates the schema if the "active" condition is not
true. But, when a driver is deactivated, we do not change the value
of the active condition. So, when it is reactivated, the schema is never
recreated.

Solution:
Delete the active condition when the driver is deactivated. Settting
the condition to falst, "", or unknown is not sufficient because it causes
the state of the resrouce in the UI to go to activating.