Set ID to name if empty and allow new settings
If request ID is empty, the name will be used instead.
If setting could not be found, it will be created. Prior,
request ID was always used to query with. Request ID is
set by parsing the request URL. It is possible for a
setting request URL to not include the setting ID, and
instead include it in the body. When this was the case,
the request would return with an error. If ID was not
found then an error was also returned. Now, the setting
can be properly set despite not having its ID in the
request URL. If a setting does not exist it will be
created.