Explicitly use int64 for owner decision
Current logic to decide owner of cluster is depending on int64.
This is the logic quoted from pkg/api/controllers/usercontrollers/usercontroller.go

> ck := crc32.ChecksumIEEE([]byte(cluster.UID))
> scaled := int(ck) * len(peers.IDs) / math.MaxUint32

We expect that int(ck) * len(peers.IDs) must be overflow of int32.
That's why we here cast int32 with int. but int behavior would
be different depending on environment.

So if we know always we want to use int64,
It's better to use explicitly int64 instead of int