Make freeipa group membership work without entrydn
Some Freeipa servers seem to not return the entrydn attribute,
which we use in our query to get all details of the groups of a
logged in user. So in such cases, we will use queries with member
attribute to get all first level groups, and the nested logic we
have for openldap to retrieve nested parent groups.