Addresses a bug in updating rolebindings for clustertemplate
If a clustertemplate was shared with a user giving "member" access, a rolebinding
was getting created with role containing "update" verb for that user. If the access
type was then switched to read-only, user still had update access. This is because
the rbac controller wasn't handling the case when there were no longer any members
for a particular access type