Fix node taints not showing
**Problem:**
The node taints won't be set if there have duplicated taints in
node template and node pool.

**Solution:**
For the new node from node pool, the taints with same key and template
will respect the one from the node pool. The other taints from the node
template will be added to newly create node.