Add default to azure-disk shared storageclass
Now adds default storageaccounttype of Standard_LRS to azure disk
storageclass if storageaccounttype is missing and kind is shared.
Prior, a storageaccounttype could be created with an empty or
missing value for storageaccounttype which would cause a PVC that
referenced it to fail during creation. This happens in kubernetes
1.12+ because it is passing the empty value to storage account
creation, which is invalid.