Handle upgrades with previously added charts
If the chart has been added previously then the url of the icon will
not be included in the iconFilename. Instead of defaulting to serving
the icons, check if they have a url for the icon in the chart. If they
do, proceed to send the url to the UI and do not serve the icon data.