Fix parsing of charts that contain hypens
Previously, if a chart contained multiple hyphens it would break the
looking up of that chart.
Now the external id is used which has a consistent format of
catalog://?catalog=<val>&template=<val>&version=<val> which allows for
predictable lookup using the value of template to create the tag.
This will not break since '&' is an invalid character to name a chart.