Update pipeline branches link to get all branches
Problem:
It takes a long time to fetch branches and pipeline files when there are plenty of branches.

Solution:
Update branches link to get all branches. So that we can reduce API calls from n+1 to 2, one for list branches and one for getting the pipeline file in a specific branch on demand.