Enqueue crtbs on project creation
Problem: Users added as cluster-owners cannot create apps in projects added to the
cluster after they have been made cluster-owners

Cause: The crtb controller lists all existing projects and creates rolebindings
in each project for the current crtb's user. So cluster-owners get access to resources
under `projectManagmentPlaneResources` for projects that exist already in the cluster when
they are being added. The same sync needs to happen on adding a new project to the cluster
as well which wasn't happening.

Fix: On project sync, enqueue clusterroletemplatebindings of that project's cluster.