Cleanup legacy globalrolebindings finalizers to deleted clusters
Previously, this only removed finalizers from deleted grbs with a 1 hour delay.
This meant that when a user had their admin access removed it would take one hour for
the change to propogate if they had a finalizer to a non-existent cluster in their list of finalizers.

Now, we premptively remove "bad" finalziers from grbs that point to non-existent clusters so that on
change they will not need to be cleaned. This only occurs once and all grbs created after this fix will
have the "field.cattle.io/grbUpgrade" annotation to avoid running through this sync handler.