Improve logic for checking logging app deploy success
Problem:
Enable logging in a PSP restricted cluster. Check that logging app is deployed but no fluentd pods are deployed due to PSP limitation, disable logging, and notice that the logging app is not removed.

Solution:
This is cause by wrong use of rancher.ticker, when the timeout, the ticker get stop, but the ticker channel isn't close cause loop

Issue:
https://github.com/rancher/rancher/issues/22797