Take 2
I'll scrub these commits before we merge.
The problem with original style was as long as we are syncing certs from the store,
we never know which ones were generated locally. And therefore, we never know which
ones we should update. This version only updates if it matches a local
key. We still have to sync later in case the original save fails.