Remove invalid parsing of distinguishedName in AD refresh principals
ExternalID is the distinguishedName of a user, which is being used as the baseDN
in ldap search query in RefetchGroupPrincipals. This doesn't need to be split on `\`.
That split is only for when we're using some field that has "logindomain\username"
format. Splitting incorrectly on `\` removed a part of the distinguishedName leading
to ldap errors