Log cert parse error instead of returning err
The secrets store calls AddCertInfo to parse cert so it populates all the necessary
UI fields. But if the cert is malformed AddCertInfo throws an error. This also
prevents the project in which this cert exists from loading in UI.
This commit logs the error message instead of returning error. This way the cert
will still not display in UI properly since it is malformed, but this won't stop
the project from loading.