Update CRB and RB if roleTemplate's inherited roles are edited
When a user is added to project/cluster, based on the roleTemplate, rolebindings
and clusterrolebindings are created for that user.

When user is added to a  project, following clusterRoleBindings are created
* roleTemplate.Name-promoted: if storageClass, persistentVolumes and apiservices are
among the roles's resources
* project.Name-namespaces-edit and create-ns: if there is "create" or "*" permission on the
"namespaces" resource of any of the rules in the roleTemplate
* project.Name-namespaces-readonly: if there is no "create" permission on "namespaces" resources
in any of the rules in roletemplate

Along with this, roleBindings are created in the namespaces belonging to that project:
* One with the clusterRole of the roleTemplate.Name
* One with the Kubernetes default role associated with the roleTemplate

The project-member role inherits the "edit" Kubernetes default role. So when a new roleTemplate
is created by cloning this role, that will also inherit the "edit" role.

The bug is that on editing such a cloned role and removing the "edit" inherited role, and
replacing it with "view", the corresponding roleBindings don't get updated.

This commit fixes it by calling the same logic that exists on the creation/edit of a prtb,
on the edit of a roleTemplate as well.