always pass node-ip parameter to kubelet
this fixes issue #25042. Where I started my rancher with the "internal-address"
and "address" both set to the ip address on eth1(192.168.33.0/24 in my case), to
make it not auto choose the ip adress eth0(10.0.2.15) which in my case was used
for NAT on vagrant. Because "address" and "internal-address" were both set to
the same value the "node-ip" parameter was never passed to the kublet, and the
kublet autoresolved its ip to the one on eth0.