Move roletemplate out of cluster lifecycle
Problem:
cluster-roletemplate-sync was cluster scoped which means with 2k clusters there will 2k annotations and 2k finalizers on each roleTemplate. This causes slowdown when listing and also makes the object impossible to look at in general.

Solution:
Move cluster-roletemplate-sync to a management level lifecycle to handle the downstream cluster deletion of the cluster roles and setup a cluster scoped handler to handle syncing the roleTemplate to clusterTemplate on changes.