Set upgrading status sooner
Problem: Previously the updating state would be set after plans were
deployed into the downstream cluster which meant that the whole helm
install would need to occur before the cluster entered upgrading.

Solution: When an upgrade is determined to be necessary, the cluster
state is updating. This is before the system-upgrade controller is
deployed.