Lowercase resource set in proxyStore
The resource name is takes from the Norman schema PluralName field.
In norman this is a plural form of the ID field which is typically
camelCase.  A Kubernetes resource field is always lowerscase. In
Rancher <=2.3 the RBAC system would do case insensitive checks
against all strings. In Rancher 2.4 the RBAC system was switch from
Norman to Steve.  Steve's implementation is much faster and more
proper and as a result does not do case insensitive checks. It is
expected that all input to the RBAC system is proper. This is the one
place in Rancher where we weird not generating a proper resource name.