Add basic support for k3os cluster drivers
When a k3s cluster is imported there is a possiblilty that it will be a
k3os cluster. In that case, k3s specific code may no longer be
applicable, ie: the cluster upgrade will require a different plan than a
normal k3s cluster.

Now, when a cluster is imported we will look for nodes with
"k3os.io/mode" labels on them to determine if the cluster is k3os.