Add EKS specific subnet validation
Previously, subnets were editable on EKS clusters which does not represent the architecture of
Kontainer Driver which does not allow for editing of subnets after the cluster has been provisioned.

This was done as a validator due to the difficulty of adding additional fields to the kontainer
engine type since this requires regenerating the protobuff files for Kontainer Driver.