Store userinfo in userattribute
This commit adds a field UserInfo on to UserAttribute CRD. The fields in
UserInfo vary per provider and can be set as a map. For github we save the
company, location, email and github ID.
The userInfo api endpoint of github already returns all these fields.
This commit adds 3 new fields to github Account type so we can get the
values for company,location and email from github api response. And saves
all these fields on the userAttribute crd object of that user.
Although userinfo endpoint returns email too, it only does so for emails
set on the profile. For other users we need to request email from user/emails
endpoint. So this commit also adds the scope "user:email" for this request.
Now on UI won't add the existing "read:org" scope either, the api will add that too