Scope redeployAgent repo check for cluster repo
appliedRepo is only set when private_registries is configured in the cluster config, else it is nil. The check for nil is always hit when the Rancher setup is configured to only use system default registry (when the registry does not require auth), so redeployAgent will always be hit in case where Rancher is configured with only a system default registry. This change scopes the repoChange check to clusters that have a private registry configured in the cluster (and appliedRepo is not nil). When system default registry changes, this will be caught by the logic in imageChange, as this uses the full repo path to the image (including the system default registry)