Renamed provider json key name
Renamed cluster's provider field's json key from "provider" to
"clusterProvider". This is because the UI computes values for
provider which are overridden when it is provided. This causes
the EKS edit UI to not properly show. The values from this
field cannot always be used in place of the computed provider.
For example, EKSv1 and EKSv2 would have the same provider
value in the backend but need to be handled differently in the
UI.