Only use the node dialer if the host is a control plane node or ready
During cluster provisioning you could end up with a node being first
in the list that is worker only and wait to be provisioned because
the control plane is still provisioning. In this situation we chose
the worker node for the node dialer which fails because the nginx
proxy does not exist yet.  Because the node dialer fails we can't
deploy the cluster agent and get stuck in a failing loop.