Fix nodegroup length check
Prior, nodegroups were only being checked for imported clusters.
This check should be performed regardless of a cluster being
imported or not as neither type will be able to deploy the rancher
agent without one. Prior, only the spec was being checked for
nodegroups, but an imported cluster may not have nodegroups as a
managed field. In this case the nodegroups should be reflected on
the upstream spec. Now, it will be evaluated if a cluster has
more than one nodegroup whether it is imported or not. Upstream
spec will be checked for nodegroups if the field is nil on spec.