Evaluate upstream only if spec nodegroups are nil
Prior, upstream spec was being checked for nodegroup length
even if nodegroups were set on spec. This caused an error
when nodegroups were finished provisioning in EKS so the
EKS Cluster Config object was in "active" phase but the
refresh handler had not updated the upstream spec since
before the nodegroups were being created.