Do not mutate EKSv1 subnets on update
Prior, if subnets were different an error would be returned.
Before v2.5 sending a request with an empty array would result
in a nil field. Now it results in an empty array. This is
causing the error to return given the same cluster body.
To prevent field updates sometimes the "noupdate" tag is used.
These were not used before because the subnets are in a dynamic
config. Now, logic has been added to the cluster store to
implement the same behavior as the "noupdate" tag which simply
does not mutate its fields value.