Add new restricted-admin role
Problem:
With local cluster now being required the admin role will have full
access to the local cluster which is not good from a security
perspective

Solution:
Add a new restricted-admin role with limited access to the local cluster
but cluster-admin access to all downstream clusters. The
restricted-admin will not be able to create anything rbac related in the
local cluster but should have access otherwise.