Adds a new label to CRB/RB created for a CRTB/PRTB
Problem: Upgrading to 2.5 adds a new label containing the CRTB/PRTB namespace and name
to the CRB/RBs created for it. Both labels have the same value "memberhsip-binding-owner".
When a CRTB/PRTB is deleted, the indexer gets all RTBs with this label value. If there are more
than one labels with this value, the CRB/RB is not deleted. On upgrade there will be two labels with
the same value, one is the existing label with UID as key and the other is the new label with ns+name

Solution: Change the label value. With this, on an upgraded setup, the controllers will only check if
the RB/CRB has a label with the new value. There will be only one such label, so on CRTB/PRTB deletion
the RB/CRB will get deleted. This is also required when migrating an upgraded setup using backup-restore-operator