Fix rancherd not setting admin on local clusters
This commit fixed an issue where local admin is not set as the cluster
owner of local cluster. This is because when we disable the default
admin, all the controller won't be able to rerun twice since all the
conditions have been set to true. Thus in reset-admin, when we first
bootstrap the password, we need to set creatorId on the annotation and
also revert the condition to be false so that controller can be re-run
to add corresponding roles/rolebindings