Tweak rancherd latest channel
Until 2.5 is released, I need the rancherd latest channel to serve up
RCs. This is due to a chicken-and-egg problem related to a rancherd blog
I'm writing.

After 2.5 is released, we'll add the excludeRegexp back in.

This change only needs made on master because this file is read by the
channel-server service running at update.rancher.io.