Evaluate proper error
Prior, the error from deploying the eks-operator was evaluated
but the condition update error was returned. The condition
update error is usually nil. This caused the controller to not
renqueue the cluster. Since the cluster was not reenqueued the
app was not redeployed unless a separate event caused the
cluster or another cluster to reenqueue. Now, the condition
error is evaluated and the original error is returned if the
condition error is nil.