Add trailing slash to helm repo url
If the helm repo url does not end with a '/' and the chart url is
relative, then the end of the helm repo path was being removed. For
example, if the helm repo was `https://helm.example.com/repo/charts` and
the chart url was `chart-v1.0.0.tgz`, then the url was being resolved to
`https://helm.example.com/repo/chart-v1.0.0.tgz`. This was incorrect.
Ensuring that the repo url ends with a '/' fixes this problem.