Allow nodes to be deleted if cant access cluster
The node controller logic was changed, which included a change to return error on node v1 delete if it wasnt a nodenotfound error. In testing it was discovered that when a cluster is inaccessible (e.g. shutdown all etcd nodes), the error "context deadline exceeded" was returned but the node never got deleted. This adds that condition to pass so it can be deleted, and the cluster can be restored using the way it worked before this change.