Make security groups optional when VPC is provided
Previously, if a user provided a VPC ID, they were also required to
provide at least one security group. However, this requirement is
extraneous. So a user is able to provide a VPC ID and not provide any
security groups.

Conversely, if a user does provide security groups, then those security
groups are already attached to a VPC. Therefore, the VPC should be
provided as well.