Add nil check for version on status
Added nil check for version on status. Prior, version could be
accessed before it was set. This would occur on clusters that were
newly provisioned or in the process of being provisioned and did
not have version set yet. Now, if the version is not set an error
will be returned so that controllers can retry until version is
available.