Remove private endpoint check when deploying agent
When an EKS cluster is created with private-only API endpoint, the user
is responsible for deploying the cluster agent. There was a check to try
to ensure that the cluster agent was not deployed by Rancher. However,
this check is not needed because, in this scenario, there are no nodes detected before the
cluster agent is deployed.