Check for cluster connection when public access disabled
If public access is disabled, it is no longer assumed that Rancher
cannot communicate with the cluster API. Instead, Rancher attempts to
contact the API. If this is successful, then Rancher can communicate
with the cluster as normal (through the API access endpoint) instead of
tunneling.

This is most important on cluster creation/import. In this case, since
Rancher can still communicate with the API endpoint, the user does not
need to run the import command that is displayed.