Changed EKS nodegroup error message
Prior, error message did not specify that managed nodegroups were
necessary. This confused users who had cluster with non-managed
nodegroups. The message also states the agent cannot be deployed
which isn't true; it is possible the agent will deploy with a
non-managed nodegroup. Now, the message has been revised to be
more accurate and a comment has been added to explain some of
the reasoning behind the requirement.