Add support for EKS launch templates
The most significant part are the changes to syncing. The syncing
algorithm more carefully crawls the cluster spec to only sync the parts
that are not nil in the config. This is useful because it allows
tracking of the Rancher managed launch template in the upstream spec,
but this will not sync this template to the spec (the launch
template will be nil because it is not provided by the user).

The rancher managed launch template ID and template versions for
managed node groups are also tracked in status for updates.