Fix k3s upgrade dev image export
Prior, using export script with TAG set to dev, master, or a
value equal to kd.RancherVersionDev would result in no k3s
upgrade or related images being added to the images text.
This is because kd.RancherVersionDev does not contain a
valid semver version. This happens to dev and master because
those values are overwritten with the value of
kd.RancherVersionDev if detected. Now, versions are used for
comparison that match the minor and major version of
kd.RancherVersionDev and will match any constraints that
could be met by any version that would also have matching
major and minor versions.