Use namespace from the object when looking for more
Problem:
When syncing helm configMaps and secrets there is a constant loop of
updates to apps causing large amounts of load

Solution:
When attempting to compare the object to other versions the namespace
being used was based off the release, this is not the namespace the
config object lives in so there would always be zero results making it
look like we were always working on the most recent object. Use the ns
from the config object to look for more.