Generate images file that includes sources
Image export now produces a text file that includes images and
which part of rancher utilizes them. Possible image sources
are any feature or system chart, rke, k3sUpgrade, core (meaning
it is need for rancher core functionality), and rancher (which
it is either the rancher server image or the agent image).
Prior, a list of images was produced but there was no annotation
included for why each image is included. This information is
useful in deducing whether an airgap setup absolutely needs an
image.