Fix cluster template rendering of AgentEnvVars
Without this commit, if AgentEnvVars is unset, it renders to YAML as an
empty slice and gets appended to the env array in an invalid way:

```
containers:
- name: cluster-register
  imagePullPolicy: IfNotPresent
  env:
...
  - name: CATTLE_K8S_MANAGED
    value: "true"
  []
  image: rancher/rancher-agent:master-head
  volumeMounts:
...
```

This causes updates to a Cluster resource to fail. This commit ensures
that the template never bothers rendering this value no matter whether
it is nil or empty.