Fix syncing controller re-queuing
In order to ensure the clusters that need syncing are reenqueued
properly, the last cluster refresh time is set any time the cluster is
updated. Also, the cluster is reenqueued even if the cluster should not
ready for a refresh. This fixes an issue where the refreshing would stop
if a cluster is updated successfully before a refresh is needed (the
cluster would enter the handler because the cluster object was updated,
the refresh time was not exceed so the handler would return without
re-queuing). This also prevents unnecessary log messages where the
cluster is creating, so the EKS handler is re-queuing the cluster and
the refresher is continually checking for upstream changes.

In addition, some logic was changed to better align with the
function/variable names (shouldRefresh was acting more like
shouldNotRefresh).