Fix clusteroperator CAcert recorder
Without this change, the behavior of the CAcert recorder for the EKS
provisioner changed when the handlers for EKS and GKE were refactored
into a single clusteroperator package (a6f0b91b), causing the cert to be
recorded on the cluster status wrongly. This change reverts the common
package to use EKS's old behavior. The GKE operator will be updated to
work in the same way.