Add k3sconfig validation
Validation has been added to prevent a k3sConfig being set on a
cluster that does not have one set already. Prior, it could be
set which is invalid because it is expected it will be
initialized by a controller when it is detected that a cluster
is a k3s cluster and the cluster is not embedded. This made it
possible to set the k3sConfig and attempt upgrade an embedded
k3s cluster, which is not possible.