Add ability to create cloud creds for any node driver
A previous PR (31548) added support for Google cloud credential.
However, this made the unsupported Google node driver active. In order
to deactivate the node driver and create the cloud credential schema for
the Google driver, a new field is added to the NodeDriverSpec:
AddCloudCredentialByDefault. If this is `true` for any node driver, then
the driver is downloaded, if necessary, and the cloud credential schema
created. This would work for any node driver, not just the built-ins.