Add Imported field to GKE upstream spec builder
Without this patch, if a GKE cluster is imported to Rancher, and then a
change from GKE is synced back to the cluster in Rancher, "imported"
will flip back to false, and the guard in gke-operator that prevents
sending a DELETE request to imported clusters that are removed from
Rancher gets thwarted. This change fixes the upstream spec builder
function in the cluster refresher to always use the config spec value
for the Imported attribute.