Handle already exists errors during startup
Problem:
In an HA setup ranchers will race to attempt to create things, including
settings and ca cert secrets. This can cause rancher to fatal and
restart the pod.

Solution:
Handle the error and either try again in the case of setting up
listeners or ignore it when creating settings.