Remove released directory for images
The rancher/charts repo was updated to hold a staged version of the
released charts in the released directory. This change extracts the
charts into the assests directory instead of the released/assets
directory. This will make the export process the same for dev and
release branches of rancher charts.