Validate actual cluster names in GKE
Currently, the validateGKEClusterName validator checks the new cluster
name against the names and locations of clusters that Rancher already
knows about, and does not account for any potential clusters that exist
in GKE but have not been registered with Rancher. If the user creates a
cluster with the same name as another cluster in the same region or zone
that isn't known to Rancher, it will start to provision it but meet a
400 response from GKE. Since the clusterName field isn't editable, the
problem isn't fixable after the fact and the only option is to delete
and recreate the cluster. This commit changes the validator to actually
query GKE for current clusters instead of just the cluster list in
Rancher. This will provide a more accurate view of the current clusters
and will allow the user to correct the mistake before provisioning has
started.