Set condition messages for "Provisioned" condition
Without this patch, the changes in how wrangler summarizes
conditions[1][2] causes the condition type, in this case "Provisioned",
to be rendered as the cluster transition message. Since this string
doesn't match the display state "Provisioning", the transition message
is displayed underneath the display state. Setting an non-empty message
on the condition makes this slightly less confusing.

[1] https://github.com/rancher/wrangler/commit/3eba78f45e7d61364efd11d73334641c17305b4c
[2] https://github.com/rancher/wrangler/commit/3533ace946aaf8cc64ae4be9ead3221c35975884