Fix docker-runtime-publish
A previous change moved the RKE2_VERSION in the environment variable
file. This caused a problem with the docker-publish-runtime drone
pipeline because this environment file is not used there.

This change moves the RKE2_VERSION back to the
package/Dockerfile.runtime and adds comments to future updaters on how
to handle updating in the future.