Add context to Ready condition while updating
When zonal GKE clusters are updating the control plane, the control
plane usually becomes unreachable. This is expected and normal
behavior[1]. However, when the HealthSyncer controller runs, it cannot
reach the API and sets ClusterConditionReady to false with an alarming
error message. This change does not eliminate or ignore the error but
adds more context to the error message to help users understand that
this is normal.

[1] https://cloud.google.com/kubernetes-engine/docs/concepts/cluster-upgrades#upgrading_manually

(cherry picked from commit 63dacabe5427053375890d401bbb39d4a0c540f0)