Add Google OAuth to proxy
The removal of the Google basic auth means that the service account
token expires after 60 minutes. This was fixed in other areas, but the
proxy still used the service account token on the cluster object instead
of getting one from the token source. Getting the token from the token
source gets a renewed token that is usable for 60 minutes.