Use additional CA for hosted providers
When the call is made to get a service account for the hosted providers,
the CA for the k8s API is used. If Rancher is running behind a proxy or
the like that uses its own CA, then this call fails.

This change will add the additional CAs to the rest config so the CAs
are used in the call to get the service account token.