Make Azure tenantId optional in cloud credential
d4f32aca added tenantId as a public credential field for the dynamically
generated azurecredentialconfig schema. The corresponding machine driver
change[1] can optionally accept a tenant ID but can fall back on its
original lookup behavior if none is provided. Without this change,
Rancher sets the new field as required and breaks backwards
compatibility. This change adds a new "optional" category to the machine
driver data to allow the credential not to be marked as required.

[1] https://github.com/rancher/machine/pull/133