Uninstall objects from charts in non-default namespaces
If the objects of a chart were installed in a non-default namespace
without the namespace being listed in the object's metadata, then the
uninstalled would seem successful but the helm is not able to find the
objects. The non-namespaced items would be removed, but any namespaced
items would be leftover.

It should be noted that the namespace provided allowed helm to find the
release secret, but would not namespace the call to delete the objects
in the release.