Handle System/Default projects for hosted Rancher
If the creatorID annotation was missing from the local cluster object,
then Rancher would skip creating the System and Default projects. After
this change, these projects will get created regardless of the
existence of the creatorID annotation.

Also, when `app.EnsureAppProjectName` was called in preparation for
installing an app, it reassigns the namespace to the project if the
namespace annotation is not the expected projectID. This change does not
reassign the cattle-system namespace in such cases. This is to attempt
to keep the cattle-system namespace in the host cluster's tenant System
project.