Check EKS cluster access in generic proxy
In the very specific case that the /meta/proxy route is being called in
order to get data about an existing EKS cluster, allow users who are
cluster owners for that cluster but not necessarily owners of the cloud
credential to call the provider API through the proxy. This is necessary
to allow these users to edit the cluster. The UI needs to pass an
additional 'clusterID' parameter in the X-API-CattleAuth-Header header.