Conceal sensitive in audit logs
Previously, when audit logging of requests was enabled, sensitive data
was exposed in the audit logs. Now, the values identified as sensitive
are concealed before printing to the audit log. These include items
ending in "[pP]assword", "[tT]oken", values of secrets, and non-public
items in DriverData in machinedriver_data.

Co-authored-by: Ricardo Weir <ricardo@rancher.com>