Run update-ca-certificates before launching Rancher
The new SLE image requires update-ca-certificates be run before
additional trusted CAs are recognized. Therefore, the command is
run prior to launch Rancher.

In addition, the location of the mounted certificate is changed to the
location specified in the docs: /etc/pki/trust/anchors/

Finally, the jailer also copies the certificate infromation to /etc/ssl
instead of /etc/ssl/certs because the Go libraries expect the certs to
be there.