Accept zone or region for GKE versions
Without this patch, /meta/gkeVersions only accepts 'zone' as a location
parameter and so only returns kubernetes versions available in a given
zone. The available versions in a zone is not necessarily the same as
the versions available for a region, even if a zone is within the
region. This change updates the /meta/gkeVersions route to accept either
'zone' or 'region' and return the appropriate list.

Example usage:

  GET /meta/gkeVersions?projectId=my-project&cloudCredentialId=cattle-global-data:cc-abcde?region=us-central1

  GET /meta/gkeVersions?projectId=my-project&cloudCredentialId=cattle-global-data:cc-abcde?zone=us-central1-f

(cherry picked from commit 2c0d42e3b7a33f194f3cc0473b350649701e4c75)