Accept zone or region for GKE versions
Without this patch, /meta/gkeVersions only accepts 'zone' as a location
parameter and so only returns kubernetes versions available in a given
zone. The available versions in a zone is not necessarily the same as
the versions available for a region, even if a zone is within the
region. This change updates the /meta/gkeVersions route to accept either
'zone' or 'region' and return the appropriate list.

Example usage:

  POST /meta/gkeVersions
  {
    "projectId": "my-project",
    "region": "us-central1",
    "credential": "<token>"
  }

  POST /meta/gkeVersions
  {
    "projectId": "my-project",
    "zone": "us-central1-f",
    "credential": "<token>"
  }