Filter by region in /meta/aksVirtualNetworks
This change optionally enables filtering by region when listing ASK
virtual networks. It is backwards compatible: if region is not set, it
defaults to the old behavior of listing all virtual networks. Also adds
"location" as an attribute returned in the list of vnets. This isn't
functionally necessary for the UI but it's nice for developers to be
able to validate the results.