Only log cause if debugging is enabled
Prior, the WrapAPIError function was being used which set the
Cause field as the underlying error causing the issue. When
the Cause field gets set, it is automatically logged. In this
case the error is very specific to the package it comes from
and can cause confusion. Specifically it has caused confusion
because the error mentions a hashed password which is being
mistakenly interpreted as referring to the token hashing
feature. Now, the WrapAPIError function is no longer being
used and the underlying error will only be logged if debug is
enabled.