Look up tenant ID in /meta/aks* if none provided
This makes providing the tenant ID in the cloud credential optional. The
tenant ID is not stored, it is looked up every time a client is created,
which is on every /meta/aks* request. This only applies to AKSv2, where
the cloud credential is used. For AKSv1, all parameters including the
tenant ID must be passed in the request, as has always been the case.

In order to avoid making repetitive requests to the Azure cloud
authorization endpoint to retrieve the tenant ID, store the fetched
tenant ID in an annotation on the cloud credential secret. The
annotation expires after an hour and will be re-fetched. Although it is
not likely to change hour to hour, in the event that it does change we
want the credential to be usable within a reasonable delay.

Also updates aks-operator which will do the same thing.