Add VMWare provisioner to override
Some of the details for provisioning are passed as environment variables
instead of as command line arguments. A few of the provisioners have
different environment variable names compared to the command line
arguments. The VMWare provisioners are added to this.

In addition, rancher-machine version is bumped to address a hanging
issue for vsphere provisioning.