Use additional CAs when getting AKS service account
Before this change, the AKS controller in Rancher would try to get the
service account by using the kubeconfig from AKS. This caused an issue
if any additional CAs are needed (like in a proxy setup).

This fix will ensure that the additional CAs are used when getting the
service account and resolve the unsigned cert errors.