Add generation field to ETCDSnapshotCreate
Previously, the UI would put a blank ETCDSnapshotCreate object on the
cluster spec when the user wanted to create an on-demand snapshot. This
caused an issue where multiple on-demand snapshots could not be created
this way because the backend would not detect any changes (the
ETCDSnapshotCreate object would remain blank).

After this change, the UI will update the generation field when a new
on-demand snapshot should be created and the backend will detect this.