Set PeerManager to run in clustered mode
Previously, the PeerManager in the wrangler context was not set. This
caused the cluster controllers to run on every node instead of in a
clustered/shard mode. The symptom of this was that too many v3.Nodes
were getting created when deploying RKE2/K3S clusters in an HA setup.

Now, the PeerManger will be set and the cluster controllers will run as
expected.