Retry create if already exists error
Sometimes create fails despite the delete call being made
prior. This is because it is in the process of being deleted.
Already exists errors were being ignored causing the caller
to receive a 201 even though the token was not created. It
was possible that the cluster auth token was created. Now,
create will be retried if an already exists error is
encountered.