Add v1.Cluster indexers to common registration
Previously, the v1.Cluster indexer was created on the management plane,
but not on any of the followers. During sharding for the user management
plane, if the management plane was not selected, then the indexer would
not exist and could not be used. This would cause things like the
snapshotbackpopulate to not function.

Now, the v1.Cluster indexer is registered with other common indexers.