Generate v3.Cluster object name on first pass
Previously, the v3.Cluster object name was deterministic for a given
v1.Cluster based on the name and namespace. This caused an issue with
caches when the a cluster is created, deleted and then created with the
same name: the v3.Cluster object would have the same and would use the
same caches.

Now the v3.Cluster object's name is generated on first pass and then
pulled from the v1.Cluster object's status afterward.